requirements.txt
pywin32; sys_platform == 'win32'
pyusb
pyftdi
pydispatcher
wxPython
requests
lxml
numpy==1.15.4
colour
pyparsing
gitpython
git+https://github.com/MCU-Innovations/eculib.git#egg=eculib>=1.0.19
